+++ examples/data-sources/clerk_jwks/data-source.tf
@@ -1,0 +1,5 @@
+data "clerk_jwks" "this" {}
+
+output "signing_key_ids" {
+  value = [for k in data.clerk_jwks.this.keys : k.kid]
+}