+++ examples/resources/clerk_domain/resource.tf
@@ -1,0 +1,4 @@
+resource "clerk_domain" "satellite" {
+  name      = "satellite.example.com"
+  proxy_url = "https://satellite.example.com/__clerk"
+}